+++ CHANGELOG.md
@@ -8,6 +8,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+* Add option to use Draw.io Desktop in headless mode
+
 ## [1.0.0] - 2021-03-24
 
 ### Added